mail to some files
